requirements.txt
openai
python-dotenv
pyttsx3
SpeechRecognition
pyaudio
requests
pvporcupine
flac
PyYAML
apscheduler
vosk
sounddevice